Add checkstyle and license header checks
