+++ checkstyle-suppressions.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+
+<!DOCTYPE suppressions PUBLIC
+     "-//Checkstyle//DTD SuppressionFilter Configuration 1.0//EN"
+     "https://checkstyle.org/dtds/suppressions_1_0.dtd">
+
+<suppressions>
+  <suppress checks="^(?!Header$)"
+             files="src/main/java/com/google/cloud/solutions/spannerddl/parser/AST.*\.java$"/>
+  <suppress checks=".*"
+             files="target/generated-sources/.*$"/>
+</suppressions>
