+++ licence_header.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Google LLC
+ * Copyright 2024 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,28 +13,3 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package com.google.cloud.solutions.spannerddl.parser;
-
-import com.google.cloud.solutions.spannerddl.diff.AstTreeUtils;
-import com.google.common.base.Joiner;
-
-public class ASTcolumns extends SimpleNode {
-
-  public ASTcolumns(int id) {
-    super(id);
-  }
-
-  public ASTcolumns(DdlParser p, int id) {
-    super(p, id);
-  }
-
-  @Override
-  public String toString() {
-    return "( "
-        + Joiner.on(", ")
-            .skipNulls()
-            .join(AstTreeUtils.getChildrenAssertType(children, ASTkey_part.class))
-        + " )";
-  }
-}
