+++ src/main/java/com/google/cloud/solutions/spannerddl/parser/ASTcolumn_length.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.solutions.spannerddl.parser;
 
-import com.google.cloud.solutions.spannerddl.diff.ASTTreeUtils;
+import com.google.cloud.solutions.spannerddl.diff.AstTreeUtils;
 
 /** Abstract Syntax Tree parser object for "column_length" token */
 public class ASTcolumn_length extends SimpleNode {
@@ -31,7 +31,7 @@ public class ASTcolumn_length extends SimpleNode {
 
   @Override
   public String toString() {
-    return ASTTreeUtils.tokensToString(this);
+    return AstTreeUtils.tokensToString(this);
   }
 
   @Override
