+++ src/main/java/com/google/cloud/solutions/spannerddl/parser/ASTcolumn_type.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.solutions.spannerddl.parser;
 
-import com.google.cloud.solutions.spannerddl.diff.ASTTreeUtils;
+import com.google.cloud.solutions.spannerddl.diff.AstTreeUtils;
 import java.util.Locale;
 
 /** Abstract Syntax Tree parser object for "column_type" token */
@@ -62,10 +62,10 @@ public class ASTcolumn_type extends SimpleNode {
       case "ARRAY":
         return "ARRAY<" + ((ASTcolumn_type) children[0]) + ">";
       case "PG": // PG.pgtype
-        return ASTTreeUtils.tokensToString(this).toUpperCase(Locale.ROOT);
+        return AstTreeUtils.tokensToString(this).toUpperCase(Locale.ROOT);
       case "STRUCT":
         if (jjtGetNumChildren() > 0) {
-          return ASTTreeUtils.tokensToString(this);
+          return AstTreeUtils.tokensToString(this);
         } else {
           return "STRUCT <>";
         }
