+++ src/main/java/com/google/cloud/solutions/spannerddl/parser/ASTpath.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.solutions.spannerddl.parser;
 
-import com.google.cloud.solutions.spannerddl.diff.ASTTreeUtils;
+import com.google.cloud.solutions.spannerddl.diff.AstTreeUtils;
 
 /** Abstract Syntax Tree parser object for "path" token */
 public class ASTpath extends SimpleNode {
@@ -31,6 +31,6 @@ public class ASTpath extends SimpleNode {
 
   @Override
   public String toString() {
-    return ASTTreeUtils.tokensToString(this, false);
+    return AstTreeUtils.tokensToString(this, false);
   }
 }
